maint: Fix warnings on OSX 10.12.4 (clang 8)
* src/Makefile.am
* src/formatter/AdapterFormatterPlayer.cpp
* src/formatter/FormatterFocusManager.cpp
* src/formatter/FormatterFocusManager.h
* src/formatter/FormatterScheduler.cpp
* src/formatter/NclExecutionObject.cpp
* src/formatter/NclFormatterDeviceRegion.cpp
* src/formatter/NclPresentationEvent.cpp
* src/isdbt-tuner/MulticastProvider.h
* src/isdbt-tuner/Tuner.cpp
* src/main.cpp
* src/mb/Display.cpp
* src/mb/SDL2ffmpeg.cpp
* src/mb/SDLSurface.cpp
* src/mb/SDLSurface.h
* src/ncl/Descriptor.cpp
* src/ncl/FocusDecoration.cpp
* src/ncl/IntervalAnchor.cpp
* src/ncl/LayoutRegion.cpp
* src/player/AVPlayer.cpp
* src/player/NewVideoPlayer.cpp
* src/player/NewVideoPlayer.h
* src/player/Player.cpp
* src/util/functions.cpp

+++ src/ncl/IntervalAnchor.cpp
@@ -24,7 +24,7 @@ using namespace ::ginga::util;
 GINGA_NCL_BEGIN
 
 // if the representation changes, update isObjectDuration method
-const double IntervalAnchor::OBJECT_DURATION = INFINITY;
+const double IntervalAnchor::OBJECT_DURATION = (double)INFINITY;
 
 IntervalAnchor::IntervalAnchor (const string &id, double begin, double end)
     : ContentAnchor (id)
